requirements.txt
pillow
pypdf
PyPDF2
numpy
matplotlib